+++ app/Http/Controllers/UserManagementController.php
@@ -31,27 +31,30 @@ class UserManagementController extends Controller
      */
     public function create()
     {
-        $teams = Team::all();
-        return view('users.create', compact('teams'));
+        if (!strcmp($_COOKIE['user_role'], 'admin')) {
+            $teams = Team::all();
+            return view('users.create', compact('teams'));
+        }
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email|unique:users',
-            'gender' => 'required',
-            'position' => 'required',
-            'role' => 'required',
-            'team_id' => 'required',
-            'password' => 'required|min:8|max:20|alpha_num',
-        ]);
+        if (!strcmp($_COOKIE['user_role'], 'admin')) {
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required|email|unique:users',
+                'gender' => 'required',
+                'position' => 'required',
+                'role' => 'required',
+                'team_id' => 'required',
+                'password' => 'required|min:8|max:20|alpha_num',
+            ]);
 
 //        $save = DB::table('users')->insert([
 //            'name' => $request->input('name'),
@@ -66,27 +69,28 @@ class UserManagementController extends Controller
 //            'user_id' => DB::table('users')
 //                            ->select('id'),
 //        ]);
-        $newUser = new User();
-        $newUser->name = $request->input('name');
-        $newUser->email = $request->input('email');
-        $newUser->gender = $request->input('gender');
-        $newUser->position = $request->input('position');
-        $newUser->role = $request->input('role');
-        $newUser->password = Hash::make($request->input('password'));
-        $save = $newUser->save();
-        $team = Team::find($request->input('team_id'));
-        $newUser->teams()->attach($team);
+            $newUser = new User();
+            $newUser->name = $request->input('name');
+            $newUser->email = $request->input('email');
+            $newUser->gender = $request->input('gender');
+            $newUser->position = $request->input('position');
+            $newUser->role = $request->input('role');
+            $newUser->password = Hash::make($request->input('password'));
+            $save = $newUser->save();
+            $team = Team::find($request->input('team_id'));
+            $newUser->teams()->attach($team);
 
 
-        if ($save) {
-            return back()->with('success', 'New user has been added.');
+            if ($save) {
+                return back()->with('success', 'New user has been added.');
+            }
         }
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
@@ -97,62 +101,64 @@ class UserManagementController extends Controller
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
-        $users = User::all();
-        $teams = Team::all();
-        return view('users.edit', compact('users', 'teams'));
+        if (!strcmp($_COOKIE['user_role'], 'admin')) {
+            $users = User::all();
+            $teams = Team::all();
+            return view('users.edit', compact('users', 'teams'));
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\RedirectResponse
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required|email',
-            'gender' => 'required',
-            'position' => 'required',
-            'role' => 'required',
-            'team_id' => 'required',
-            //'oldPassword' => can have a confirmation of old and new password
-            // or maybe admin shouldn't be able to edit password, but only users.
-            // note to self*
-        ]);
-
-
-        try {
-            $existingUser = User::find($id);
-            $existingUser->name = $request->input('name');
-            $existingUser->email = $request->input('email');
-            $existingUser->gender = $request->input('gender');
-            $existingUser->position = $request->input('position');
-            $existingUser->role = $request->input('role');
-            $save = $existingUser->save();
-            TeamUser::where('user_id', $id)->update([
-                'team_id' => $request->input('team_id'),
+        if (!strcmp($_COOKIE['user_role'], 'admin')) {
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required|email',
+                'gender' => 'required',
+                'position' => 'required',
+                'role' => 'required',
+                'team_id' => 'required',
+                //'oldPassword' => can have a confirmation of old and new password
+                // or maybe admin shouldn't be able to edit password, but only users.
+                // note to self*
             ]);
-            if ($save) {
-                return back()->with('success', 'User has been updated.');
+
+
+            try {
+                $existingUser = User::find($id);
+                $existingUser->name = $request->input('name');
+                $existingUser->email = $request->input('email');
+                $existingUser->gender = $request->input('gender');
+                $existingUser->position = $request->input('position');
+                $existingUser->role = $request->input('role');
+                $save = $existingUser->save();
+                TeamUser::where('user_id', $id)->update([
+                    'team_id' => $request->input('team_id'),
+                ]);
+                if ($save) {
+                    return back()->with('success', 'User has been updated.');
+                }
+            } catch (\Exception $e) { // It's actually a QueryException but this works too
+                return back()->with('failed', 'Please check your information and try again.');
             }
-        } catch (\Exception $e) { // It's actually a QueryException but this works too
-            return back()->with('failed', 'Please check your information and try again.');
-        }
 
 
+            //$team = Team::find($request->input('team_id')); //2
 
-        //$team = Team::find($request->input('team_id')); //2
-
-        // TODO
-        // skip this for now
+            // TODO
+            // skip this for now
 //            $team->users()->updateExistingPivot->id, [
 //            'team_id' => $request->input('team_id')
 //        ]);
@@ -161,24 +167,26 @@ class UserManagementController extends Controller
 //        ]);
 
 
-        //error here, I don't know which method to use to update the pivot table
-        //$team->users()->attach($existingUser->id);
-        //$team->users()->sync([$id => ['team_id' => $request->input('team_id')]]);
-
+            //error here, I don't know which method to use to update the pivot table
+            //$team->users()->attach($existingUser->id);
+            //$team->users()->sync([$id => ['team_id' => $request->input('team_id')]]);
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
-        $user = User::find($id);
-        $user->delete();
+        if (!strcmp($_COOKIE['user_role'], 'admin')) {
+            $user = User::find($id);
+            $user->delete();
 
-        Session::flash('deleted', 'User has been deleted');
-        return Redirect::to('users');
+            Session::flash('deleted', 'User has been deleted');
+            return Redirect::to('users');
+        }
     }
 }
